Use actual gem bin script rather than custom code to execute rake.
We can be sure it's the right rake by using the -S option to ruby.
To make this work, we need to use $@ rather than $* to pass through the
arguments untouched.

Also, a `jruby` command is required for rake: the shebang line for rake
on JRuby refers to it directly.