Stop trying to update a tag to mark the latest green build
This fixes the issue wherein developers' local copies of the "latest green
build" marker wouldn't be updated after pulling even when a CI machine had
updated the reference.

See "DISCUSSION On Re-tagging" in the git-tag(1) man page for details about why
updating tags intentionally causes downstream-developer pain.

(We're not naming the new branch the same thing as the tag so that it's not
ambiguous -- that is, so people don't have to type refs/heads and refs/tags all
the time. Also, "latest green build" is what people actually call it!)