Clean bundle after install
This allows you to cache your bundle between builds to speed up jetpack.
The downside of caching is that you can end up with a lot of stale gems
in the bundle, which can negatively affect app load time. With the
--clean option, only the currently used gems are left in the bundle.

Performance impact of this change for non-cached bundles should be
minimal. In my testing there was no perceivable difference in run times
for bundle install with an unnecessary --clean.